+++ docs/changelog.rst
@@ -63,6 +63,8 @@ Detailed list of changes
 
 - Add an option :opt:`background_tint_gaps` to control background image tinting for window gaps (:iss:`5596`)
 
+- A new option :opt:`undercurl_style` to control the rendering of undercurls (:pull:`5883`)
+
 - Bash integration: Fix ``clone-in-kitty`` not working on bash >= 5.2 if an environment variable values contain newlines or other special characters (:iss:`5629`)
 
 - A new :ac:`sleep` action useful in combine based mappings to make kitty sleep before executing the next action